chore: move serializers into ics20-transfer/types
