+++ crates/ibc-apps/ics20-transfer/types/src/amount.rs
@@ -14,7 +14,7 @@ use super::error::TokenTransferError;
 #[derive(Copy, Clone, Debug, Eq, PartialEq, PartialOrd, Ord, Display, From, Into)]
 pub struct Amount(
     #[cfg_attr(feature = "schema", schemars(with = "String"))]
-    #[serde(serialize_with = "ibc::serializers::serde_string::serialize")]
+    #[serde(serialize_with = "crate::serializers::serialize")]
     #[serde(deserialize_with = "deserialize")]
     U256,
 );
