+++ crates/ibc-apps/ics20-transfer/types/src/lib.rs
@@ -34,6 +34,9 @@ pub mod msgs;
 #[cfg(feature = "serde")]
 pub mod packet;
 
+#[cfg(feature = "serde")]
+pub(crate) mod serializers;
+
 pub mod error;
 mod memo;
 pub use memo::*;
